+++ src/main/java/com/example/bookingmeeting_be/services/DeviceService.java
@@ -12,14 +12,20 @@ public class DeviceService {
     @Autowired
     private DeviceRepository repository;
 
-    public List<Device> getAll() { return repository.findAll(); }
-    public Device create(Device device) { return repository.save(device); }
-    public Device update(Long id, Device device) {
+    public List<Device> getAll() {
+        return repository.findAll();
+    }
+    public Device create(Device device) {
+        return repository.save(device);
+    }
+    public Device update(int id, Device device) {
         device.setId(id);
         return repository.save(device);
     }
-    public void delete(Long id) { repository.deleteById(id); }
-    public void updateQuantity(Long id, int delta) {
+    public void delete(int id) {
+        repository.deleteById(id);
+    }
+    public void updateQuantity(int id, int delta) {
         Device d = repository.findById(id).orElseThrow();
         d.setQuantity(d.getQuantity() + delta);
         repository.save(d);
